[MOVE] Moved code to ol_stock module
[MOVE] Moved code to ol_purchase module
