+++ fulton_accounting_enhance/__init__.py
@@ -1,1 +1,2 @@
 from . import models
+from . import reports
