+++ ls_pdf_reports/__manifest__.py
@@ -30,19 +30,12 @@
         'security/ir.model.access.csv',
         'data/generic_paperformats.xml',
         # 'reports/invoice_rma_refund.xml',
-        'reports/packing_slip.xml',
-        'reports/packing_slip_from_so.xml',
-        'reports/picking_list.xml',
-        'reports/picking_list_from_so.xml',
         # TODO: "helpdesk.rma" model not found
         # 'reports/rma_repair_summary.xml',
         # 'reports/rma_bin_label.xml',
         'reports/mo_bin_label.xml',
-        'reports/inventory_sheets.xml',
         'reports/automation_test.xml',
         'reports/purchase_order_doc.xml',
-        'reports/product_labels.xml',
-        'reports/inventory_label.xml',
         'data/purchase_mail_data.xml',
         'views/report_extra_content.xml',
     ],
