+++ ls_pdf_reports/models/__init__.py
@@ -1,9 +1,7 @@
 from . import tools
-from . import stock_picking
 from . import mrp
 # TODO: "helpdesk.rma" model not found
 # from . import rma
 from . import purchase
-from . import product
 from . import report_extra_content
 from . import report
