+++ ls_pdf_reports/reports/__init__.py
@@ -1,8 +1,3 @@
 # from . import invoice_rma_refund
-from . import packing_slip
-from . import packing_slip_from_so
-from . import picking_list
-from . import picking_list_from_so
 from . import rma_repair_summary
-from . import inventory_sheets
 from . import automation_test
