+++ ol_stock/__manifest__.py
@@ -19,5 +19,11 @@
         "data/stock_location_data.xml",
         "security/security_group.xml",
         "views/stock_picking_type_views.xml",
+        'reports/inventory_label.xml',
+        'reports/inventory_sheets.xml',
+        'reports/packing_slip_from_so.xml',
+        'reports/packing_slip.xml',
+        'reports/picking_list_from_so.xml',
+        'reports/picking_list.xml',
     ],
 }
