+++ ol_stock/reports/__init__.py
@@ -1,0 +1,5 @@
+from . import packing_slip
+from . import packing_slip_from_so
+from . import picking_list
+from . import picking_list_from_so
+from . import inventory_sheets
