+++ ol_stock/reports/inventory_label.xml
@@ -27,8 +27,8 @@
     <field name="model">product.product</field>
     <field name="report_type">qweb-pdf</field>
     <field name="print_report_name">'%s Inventory Label' % (object.name or '')</field>
-    <field name="report_name">ls_pdf_reports.report_stock_inventory_label</field>
-    <field name="report_file">ls_pdf_reports.report_stock_inventory_label</field>
+    <field name="report_name">ol_stock.report_stock_inventory_label</field>
+    <field name="report_file">ol_stock.report_stock_inventory_label</field>
     <field name="binding_model_id" ref="product.model_product_product"/>
     <field name="paperformat_id" ref="product_inventory_paperformat" />
   </record>
