+++ ol_stock/reports/inventory_sheets.xml
@@ -7,8 +7,8 @@
     <field name="print_report_name">'Inventory Sheets'</field>
     <field name="model">stock.inventory</field>
     <field name="report_type">qweb-pdf</field>
-    <field name="report_name">ls_pdf_reports.report_inventory_sheets</field>
-    <field name="report_file">ls_pdf_reports.report_inventory_sheets</field>
+    <field name="report_name">ol_stock.report_inventory_sheets</field>
+    <field name="report_file">ol_stock.report_inventory_sheets</field>
     <field name="binding_model_id" ref="stock_inventory.model_stock_inventory"/>
     <field name="paperformat_xmlid_base">ls_templates.ls_report_paper_format_landscape</field>
   </record>
@@ -19,7 +19,7 @@
       <t t-foreach="docs" t-as="doc">
         <t t-set="inventory" t-value="doc"/>
         <!-- Call the template document with the language set -->
-        <t t-call="ls_pdf_reports.report_inventory_sheets_document"/>
+        <t t-call="ol_stock.report_inventory_sheets_document"/>
       </t>
     </t>
   </template>
